backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0

# AI/ML - Gemini
google-generativeai==0.3.0

# Agentic AI Frameworks
crewai==0.203.0
langchain==0.1.13
langchain-google-genai==0.0.9
langgraph==0.0.18
pydantic-ai==0.5.0
anthropic==0.21.0

# Database - MySQL
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Data Science & ML
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4

# API and Validation
pydantic==2.5.2
pydantic[email]==2.5.2
httpx==0.25.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# Authentication & Security
bcrypt==4.1.1
pyjwt==2.8.0

# Geolocation & Maps
geopy==2.4.1
folium==0.15.0

# Email & Notifications
sendgrid==6.11.0
twilio==8.10.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production
gunicorn==21.2.0